+++ src/main/java/com/dsmpear/main/payload/response/UserListResponse.java
@@ -11,7 +11,7 @@ public class UserListResponse extends PageResponse {
     private List<UserResponse> userResponses;
 
     @Builder
-    public UserListResponse(int totalElements, int totalPages, List<UserResponse> userResponses) {
+    public UserListResponse(Long totalElements, int totalPages, List<UserResponse> userResponses) {
         super(totalElements, totalPages);
         this.userResponses = userResponses;
     }
